use switch instead of nested ifs
This indents the code less. There is no performance penalty.

Also remove the LastPosition constants, as these are directly
related to the SliceSize constants, and it's easier to understand
if those are used instead IMHO.

Also make the go.mod file reflect the actual dependencies and check in the go.sum file.